Refactor journal creation urls
This should make the urls for journal creation easier to work with
during automated tasks. Specifically, the create and new actions for
the journal controller have been more tightly tied to a publisher. In
order to avoid going too deep into nested resources, journals are still
a top level resource, while the new and create actions are manually
mapped to the controller.