Add query slop parameter to search
The dismax query parser turns a search for foo-bar into a phrase query
"foo bar". In combination with the EdgeNGramFilterFactory this will
not return any results in most cases. I can't find any way to change
the behavior of the dismax query parser, so I'm adding a query slop
param. Given that we're just searching titles this seems like it
should fix the problem cases without diluting the results.